using size_t, fixed conversion warnings
* size_t for indizes
* -Wconversion
* hide ncurses window int indizes

+++ src/board_window.hpp
@@ -14,11 +14,11 @@ public:
     BoardWindow(int starty, int startx, Board& board);
     ~BoardWindow();
     void draw_grid();
-    std::pair<int, int> idx2win(int row, int col) const;
+    std::pair<int, int> idx2win(size_t row, size_t col) const;
     int get_height() const;
     int get_width() const;
 
-    char get_char(int row, int col) const;
+    char get_char(size_t row, size_t col) const;
 
     void cursor_up();
     void cursor_down();
@@ -26,13 +26,13 @@ public:
     void cursor_right();
 
     bool cursor_toggle();
-    void cursor_update(int row, int col);
-    void cursor_draw(int row, int col);
-    void cursor_remove(int row, int col);
+    void cursor_update(size_t row, size_t col);
+    void cursor_draw(size_t row, size_t col);
+    void cursor_remove(size_t row, size_t col);
     void cursor_show();
     void cursor_hide();
-    void update_position(int row, int col);
-    void toggle_position(int row, int col);
+    void update_position(size_t row, size_t col);
+    void toggle_position(size_t row, size_t col);
 
     coords_t select_position();
     void place_ships();
@@ -43,8 +43,8 @@ public:
     static const int cell_height = 2;
 private:
     Board& board_;
-    int cursor_y_;
-    int cursor_x_;
+    size_t cursor_y_;
+    size_t cursor_x_;
     bool display_cursor_;
 };
